requirement.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-phone-field==1.8.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
phonenumbers==8.13.27
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
